# docs/protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Super TTS daemon protocol",
    "description": "HTTP/1.1 + JSON over a Unix domain socket at `$XDG_RUNTIME_DIR/tts/super-tts-http.sock` (override with `SUPER_TTS_HTTP_SOCKET`). There is no TCP listener: the socket's filesystem permissions are the first layer of access control, and the daemon reads `SO_PEERCRED` on each connection to identify the calling binary.\n\nEvery endpoint except `POST /v1/auth/request` requires `Authorization: Bearer <token>`. A token is minted only after the user approves your app in a consent popup, and is bound to the approved binary — an app cannot widen its own permissions. See `docs/protocol/auth.md`.\n\nBecause the transport is a Unix socket, the `servers` entry below is nominal; point your client at the socket and use any `Host`. With curl:\n\n```\ncurl --unix-socket \"$XDG_RUNTIME_DIR/tts/super-tts-http.sock\" \\\n     -H \"Authorization: Bearer $TTS_TOKEN\" \\\n     http://tts.local/v1/ping\n```",
    "contact": {
      "name": "Super TTS",
      "url": "https://github.com/jorge-menjivar/super-tts"
    },
    "license": {
      "name": "GPL-3.0-only",
      "identifier": "GPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://tts.local",
      "description": "Nominal host; the transport is the Unix socket"
    }
  ],
  "paths": {
...
    "/v1/backend/list": {
      "get": {
        "tags": [
          "backends"
        ],
        "summary": "List installed backends",
        "description": "Every backend installed on this machine, each with the models it serves, the options it exposes, and which of its secrets are configured. Secret *values* are never returned — only whether each is set.\n\nThis is the full catalog. `GET /pipeline/{stage}/model/list` is the narrower read a stage's model picker wants, and browsing what is *available to install* is `GET /registry/backend/list`.",
        "operationId": "list_backends",
        "responses": {
          "200": {
            "description": "The installed catalog.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackendCatalog"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/gpu_info": {
      "get": {
        "tags": [
          "hardware"
        ],
        "summary": "Inventory the host's GPUs",
        "description": "What the daemon can see of this machine's accelerators: one entry per detected GPU with its memory, plus the host-wide driver and runtime versions that decide which backend builds will actually run here.\n\nDetection is a live probe, so this reflects the machine now rather than a cached answer. A host with no GPU answers `200` with an empty list.\n\nThis is the machine, not a model. To find out where one model may actually be placed — the intersection of this inventory with the builds that model ships — ask `GET /pipeline/{stage}/model/{model}/device/list`.",
        "operationId": "get_gpu_info",
        "responses": {
          "200": {
            "description": "The detected GPUs and host toolchain versions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GpuInventory"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/pipeline/{stage}/backend/list": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "List the backends that can fill a stage",
        "description": "The installed backends serving at least one model this stage can run, in the shape `GET /backend/list` returns them — each with its models, voices, options and secrets.\n\nFill a stage's backend picker from this rather than from `GET /backend/list`: a backend serving nothing this stage can run is refused by `POST /pipeline/{stage}`, and offering one hands the user an error to discover by choosing it. The two lists hold the same backends while `synthesis` is the only role, so the choice costs nothing today and keeps a client correct once it is not.\n\nEmpty when nothing installed serves this stage — the state that should read as \"install a backend\" rather than as an empty dropdown. `GET /registry/backend/list` is where a client sends the user next.",
        "operationId": "list_stage_backends",
        "parameters": [
          {
            "name": "stage",
            "in": "path",
            "description": "Pipeline position. `1` synthesizes — text in, audio out — and is the only position this build has; any other is a `404 unknown_stage` naming the ones that do exist.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The backends on offer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackendCatalog"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such stage (`unknown_stage`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
    "/v1/pipeline/{stage}/device/list": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "List the devices this stage's backend can run on",
        "description": "The devices the backend filling this stage can run on this host, without naming a model: the union of the per-model lists over the models it serves in this stage's role, always in `cpu`, `gpu` order.\n\nUse it before a model is chosen. Once one is, `GET /pipeline/{stage}/model/{model}/device/list` is the narrower and more accurate answer — a backend whose large model needs a GPU still offers `cpu` here because another of its models runs there.",
        "operationId": "list_stage_devices",
        "parameters": [
          {
            "name": "stage",
            "in": "path",
            "description": "Pipeline position. `1` synthesizes — text in, audio out — and is the only position this build has; any other is a `404 unknown_stage` naming the ones that do exist.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The devices on offer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceList"
                }
              }
            }
          },
          "400": {
            "description": "The stage has no backend selected (`invalid_backend`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such stage (`unknown_stage`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/pipeline/{stage}/model/list": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "List the models a stage can run",
        "description": "The models the backend filling this stage serves *in this stage's role* — synthesis models for stage 1 — as `[name, source]` pairs, which is the pair `POST /pipeline/{stage}/model` accepts. Fill a model picker from this.\n\nScoped twice over, and both halves matter. A model from another backend cannot load here without also changing the stage's backend, so offering it hands the user a pick that does something other than what the list implied. The role filter is invisible today, since every model a TTS backend serves synthesizes; it stops being invisible the moment a second position exists, and doing it in the daemon rather than in each client is what keeps the offer and the acceptance from drifting.\n\nThe full catalog — every installed backend, with the voices, language tags and device support of each — is `GET /backend/list`. This is the narrow read a stage's picker wants.",
        "operationId": "list_stage_models",
        "parameters": [
          {
            "name": "stage",
            "in": "path",
            "description": "Pipeline position. `1` synthesizes — text in, audio out — and is the only position this build has; any other is a `404 unknown_stage` naming the ones that do exist.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The models on offer, `[name, source]` pairs. Empty when the stage has no backend selected, which reads correctly as \"choose a backend first\".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelList"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such stage (`unknown_stage`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/pipeline/{stage}/model/{model}/device/list": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "List the devices a model can run on here",
        "description": "What this machine can actually offer this model: the devices its manifest declares, narrowed to the accelerators its installed asset ships and to what the host has. A CUDA-only backend on a machine with no NVIDIA GPU installed its CPU asset, and is not offered a GPU here.\n\nFill a device picker from this rather than from `GET /gpu_info`, which reports the hardware without regard to what a model supports — that is the memory figure a large model's requirements are weighed against, not the source of a picker.\n\nEmpty for a model that runs remotely, and for a local model this install cannot run on any device. A client hides the device control on an empty list rather than special-casing a status.",
        "operationId": "list_model_devices",
        "parameters": [
          {
            "name": "stage",
            "in": "path",
            "description": "Pipeline position. `1` synthesizes — text in, audio out — and is the only position this build has; any other is a `404 unknown_stage` naming the ones that do exist.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1
          },
          {
            "name": "model",
            "in": "path",
            "description": "The model's name, as `GET /pipeline/{stage}/model/list` or `GET /backend/list` spells it. Resolved against the backend filling this stage.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The devices on offer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceList"
                }
              }
            }
          },
          "400": {
            "description": "The stage has no backend selected, so there is nothing to resolve `model` against (`invalid_backend`), or that backend serves no such model (`invalid_model`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such stage (`unknown_stage`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackendCatalog": {
        "type": "object",
        "description": "Every installed backend, with its models, options and secrets.",
        "required": [
          "status",
          "backends"
        ],
        "properties": {
          "backends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BackendInfo"
            }
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "description": "The error envelope, as [`transport.md`] specifies it.\n\n`error_code` is the stable, machine-readable identifier clients switch on\nand the field that determines the HTTP status; `message` is prose for a\nhuman and may be reworded at any time.\n\n[`transport.md`]: https://github.com/jorge-menjivar/super-tts/blob/main/docs/protocol/transport.md",
        "required": [
          "status"
        ],
        "properties": {
          "error_code": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorCode",
                "description": "The stable identifier for this failure. Absent only on an unclassified\nserver-side error."
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "Reason": {
        "type": "object",
        "description": "The `data` object of a [`ReasonEnvelope`].",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Which case of `message` occurred — e.g. `expired`, `exe_changed`,\n`user_denied`."
          }
        }
      },
      "ReasonEnvelope": {
        "type": "object",
        "description": "The auth-failure envelope: `message` names the failure and `data.reason`\nsays which of its cases occurred.\n\nDistinct from [`ErrorEnvelope`] because the auth surface predates\n`error_code` and clients read `data.reason` there. Both are documented\nrather than reconciled, since changing either is a breaking wire change.",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Reason"
          },
          "message": {
            "type": "string",
            "description": "The failure identifier, e.g. `invalid_session` or `auth_denied`."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from `POST /v1/auth/request`. Bound to the calling binary and valid for 30 days."
      }
    }
  },
...
}